+++ README.md
@@ -1,8 +1,8 @@
 # java-chess
 ### Todo
 * 전체적으로 Refactor
-* Pawn 대각선 이동일 때 상대팀 말 있어야 이동 가능
-    * 상대팀 말 없으면 이동 불가
+* Move 중복 제거
+* 수학 관련 메소드 분리
 
 ### Done
 * Board 초기화
@@ -22,3 +22,5 @@
 * 죽은 말이 King인지 체크
 * 점수 계산 (status)
     * Pawn 점수 0.5점 체크
+* Pawn 대각선 이동일 때 상대팀 말 있어야 이동 가능
+    * 상대팀 말 없으면 이동 불가
