+++ src/main/java/chess/domain/PieceFactory.java
@@ -1,5 +1,7 @@
 package chess.domain;
 
+import chess.domain.pieces.*;
+
 import java.util.Arrays;
 import java.util.List;
