+++ src/main/java/chess/domain/pieces/Bishop.java
@@ -1,4 +1,7 @@
-package chess.domain;
+package chess.domain.pieces;
+
+import chess.domain.Team;
+import chess.domain.Type;
 
 public class Bishop extends Piece {
     public Bishop(Team team) {
