+++ src/main/java/chess/domain/pieces/Blank.java
@@ -1,4 +1,7 @@
-package chess.domain;
+package chess.domain.pieces;
+
+import chess.domain.Team;
+import chess.domain.Type;
 
 public class Blank extends Piece {
     public Blank() {
