+++ src/main/java/chess/domain/pieces/Knight.java
@@ -1,4 +1,9 @@
-package chess.domain;
+package chess.domain.pieces;
+
+import chess.domain.Direction;
+import chess.domain.Point;
+import chess.domain.Team;
+import chess.domain.Type;
 
 import java.util.ArrayList;
 import java.util.List;
