+++ src/main/java/chess/domain/pieces/Queen.java
@@ -1,4 +1,7 @@
-package chess.domain;
+package chess.domain.pieces;
+
+import chess.domain.Team;
+import chess.domain.Type;
 
 public class Queen extends Piece {
     public Queen(Team team) {
