+++ src/main/java/chess/domain/pieces/Rook.java
@@ -1,4 +1,7 @@
-package chess.domain;
+package chess.domain.pieces;
+
+import chess.domain.Team;
+import chess.domain.Type;
 
 public class Rook extends Piece {
     public Rook(Team team) {
